deep_research_app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
exa-py
cerebras-cloud-sdk
pydantic
